requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
deprecation==2.1.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
frozenlist==1.4.1
gotrue==2.9.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
multidict==6.1.0
packaging==24.1
postgrest==0.17.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
typing_extensions==4.12.2
websockets==13.1
yarl==1.15.2
djangorestframework-simplejwt==5.3.1
bcrypt==4.2.0
django-allauth==65.2.0
requests==2.32.3
cryptography==43.0.3
google-auth==2.36.0